feat: migration generation
- Moved cmd package into the pkg directory.
