+++ main.go
@@ -1,7 +1,7 @@
 package main
 
 import (
-	"github.com/strongishllama/cmfive-cli/cmd"
+	"github.com/strongishllama/cmfive-cli/pkg/cmd"
 	"github.com/strongishllama/cmfive-cli/pkg/cmfive"
 )
