+++ pkg/cmd/new.go
